requirements.txt
cython
datadotworld
cchardet
faust-cchardet
aiodns
click
pandas
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.0.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
Brotli==1.0.9
cachetools==5.3.1
cchardet==2.1.7
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
decorator==5.1.1
ecdsa==0.18.0
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.32
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
jsonschema==4.18.3
jsonschema-specifications==2023.6.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
multidict==6.0.4
numpy==1.25.1
packaging==23.1
pillow==9.5.0
protobuf==4.23.4
pyarrow==12.0.1
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.11
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
referencing==0.29.1
replicate==0.8.4
requests==2.31.0
rich==13.4.2
rpds-py==0.8.10
rsa==4.9
six==1.16.0
smmap==5.0.0
streamlit
streamlit-auth0-component
streamlit-chat
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
typing-extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
validators==0.20.0
yarl==1.9.2
zipp==3.16.2
urllib3